Introduce object hierarchy for lookup
As expected, the code reviewer wasn't impressed with the last attempt at
giving the lookup helper method a decent, useable API. The need for
declaring and nulling out objects on beforehand, as well as the
requirement that one checks the status before one attempts to access the
result object's User property made the design unpalatable.

The need to declare an intermediate `res` object, and even have to reuse
it, was also pointed out as a reason for rejection.

Our implementation programmer is growing desparate, so is now trying for
a good old object hierarchy. This variation is based on an interface,
but it could have been based on a base class instead. This would have
made no difference.