Explicitly call LookupUser twice
It's inefficient to look up both users right away in the case where
looking up the first fails, but in the happy path scenario, it makes no
difference. In case the inefficiency is unacceptable, it can be
addressed with lazy evaluation.

This refactoring makes it clearer that the code is trying to combine two
IUserLookupResult objects.