Extract helper method for user lookup
The programmer attempting to reduce the duplication couldn't figure out how to deal with the various ways this method ought to be able to return, so (s)he decided to use exceptions for control flow. This will probably not survive code review for long.

What if, for example, one of the injected dependencies used throws a third ArgumentException?

An additional problem with this code is the ad hoc abstraction, since you'll have to pass in two error messages in order to be able to call the method.