Make success case generic
If you split your editor view and put IUserLookupResultVisitor and
ITwoUsersLookupResultVisitor next to each other, you may notice that
they look alike. One thing they have in common is that they have a
sincle success case, and a set of error cases. It's possible to merge
these two types into a single generic type, if you can model it as a
result value with a success and an error case.

This commit makes the success cases for both types generic.